pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "esign-employee-name"
version = "0.1.0"
description = "PDF e-signature field placement at 'Employee Name' text locations using pyHanko"
requires-python = ">=3.10"
dependencies = [
    "pyhanko>=0.25",
    "pdfminer.six>=20231228",
    "pypdf>=4.0",
    "typer>=0.12",
    "fastapi>=0.115",
    "uvicorn>=0.30",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "fpdf2>=2.8",
]

[project.scripts]
esign = "esign.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/esign"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
